Fix high severity defects found in code review
readTrack() returns int and signals failure with -1, but three call sites stored
the result in a uint8_t. The value truncates to 255, so "errors != -1" was always
true and the read-failed branch was unreachable. XCopy::sendBlock() consequently
shipped garbage sector data, checksums and histogram to the web UI after a failed
read, and the "read" and "bootf" console commands reported failures as successes.

XCopy::onWebCommand's copyEmptyBlocks handler wrote into a 220 byte stack array
with no bound on the index, so a websocket message with more than 220 comma
separated fields overran the stack. The array was also uninitialised while the
loops below it and writeBlocksToFile() always read all 220 entries, so a short
message operated on stack garbage. Bound the write and zero the array.

XCopyDebug used sprintf(buf, "%-Ns", buf) to left justify in place. Passing the
same buffer as source and destination is undefined behaviour, and "%-20s" emits
20 characters plus a terminator into a 20 byte buffer. Replaced with a
printPadded() helper that pads on the way to the stream, which removes the
overlap, the overflow, and a 256 byte scratch write.

Verified: teensy31 builds clean.

+++ src/XCopy/XCopyCommand.cpp
@@ -276,7 +276,8 @@ void XCopyCommandLine::doCommand(String command)
     {
         Log.printf("Reading Track %2d:\r\n", param.toInt());
         gotoLogicTrack(param.toInt());
-        uint8_t errors = readTrack(false);
+        // int, not uint8_t: readTrack() signals failure with -1.
+        int errors = readTrack(false);
         if (errors != -1)
         {
             Log << F("Sectors found: ") << getSectorCnt() << F(" Errors found: ");
@@ -351,7 +352,8 @@ void XCopyCommandLine::doCommand(String command)
         else
         {
             gotoLogicTrack(0);
-            uint8_t errors = readTrack(false);
+            // int, not uint8_t: readTrack() signals failure with -1.
+            int errors = readTrack(false);
             if (errors != -1)
             {
                 Log << F("Sectors found: ") << getSectorCnt() << F(" Errors found: ");
